refactor: make validation utils a function
